# Optimize BitOps